+++ src/chain/store/chain_store.rs
@@ -252,58 +252,10 @@ where
         let _did_work = file.remove(cid);
     }
 
-    // FIXME: This function doesn't use the chain store at all.
-    //        Tracking issue: https://github.com/ChainSafe/forest/issues/3208
-    /// Retrieves block messages to be passed through the VM.
-    ///
-    /// It removes duplicate messages which appear in multiple blocks.
-    pub fn block_msgs_for_tipset(db: DB, ts: &Tipset) -> Result<Vec<BlockMessages>, Error> {
-        let mut applied = HashMap::new();
-        let mut select_msg = |m: ChainMessage| -> Option<ChainMessage> {
-            // The first match for a sender is guaranteed to have correct nonce
-            // the block isn't valid otherwise.
-            let entry = applied.entry(m.from()).or_insert_with(|| m.sequence());
-
-            if *entry != m.sequence() {
-                return None;
-            }
-
-            *entry += 1;
-            Some(m)
-        };
-
-        ts.blocks()
-            .iter()
-            .map(|b| {
-                let (usm, sm) = block_messages(&db, b)?;
-
-                let mut messages = Vec::with_capacity(usm.len() + sm.len());
-                messages.extend(
-                    usm.into_iter()
-                        .filter_map(|m| select_msg(ChainMessage::Unsigned(m))),
-                );
-                messages.extend(
-                    sm.into_iter()
-                        .filter_map(|m| select_msg(ChainMessage::Signed(m))),
-                );
-
-                Ok(BlockMessages {
-                    miner: *b.miner_address(),
-                    messages,
-                    win_count: b
-                        .election_proof()
-                        .as_ref()
-                        .map(|e| e.win_count)
-                        .unwrap_or_default(),
-                })
-            })
-            .collect()
-    }
-
     /// Retrieves ordered valid messages from a `Tipset`. This will only include
     /// messages that will be passed through the VM.
     pub fn messages_for_tipset(&self, ts: &Tipset) -> Result<Vec<ChainMessage>, Error> {
-        let bmsgs = ChainStore::block_msgs_for_tipset(&self.db, ts)?;
+        let bmsgs = BlockMessages::for_tipset(&self.db, ts)?;
         Ok(bmsgs.into_iter().flat_map(|bm| bm.messages).collect())
     }
